stm32/hrtim: move advanced_pwm to module
